# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Crowdfunding Platform API
  version: 0.3.0
  description: API cho nền tảng gây quỹ cộng đồng Góp Mầm.
servers:
  - url: http://localhost:4000/api/v1
    description: Backend local
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: Danh sách người dùng (chỉ admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Danh sách người dùng
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /campaigns:
    get:
      operationId: listCampaigns
      summary: Danh sách chiến dịch công khai
      description: >-
        Không cần đăng nhập. Mặc định chỉ trả chiến dịch ở trạng thái công
        khai (`approved`, `active`, `success`, `ended`) trừ khi lọc `status`.
        Lọc `status` chỉ nhận trạng thái đã công khai; draft/pending/needs_info/rejected/cancelled trả 400.
        Mọi bộ lọc chạy ở server (truy vấn lấy id rồi mới nạp quan hệ nên phân trang đúng).
        `q`/`location` thoát ký tự `%`/`_` của LIKE. `sort=progress` theo tỷ lệ
        `currentAmount/goalAmount`; `sort=ending` bỏ chiến dịch đã quá hạn. Khoảng ngược
        (`minGoal > maxGoal`, `minProgress > maxProgress`) trả 400.
      parameters:
        - $ref: "#/components/parameters/CampaignQ"
        - $ref: "#/components/parameters/CampaignCategory"
        - $ref: "#/components/parameters/CampaignStatusFilter"
        - name: location
          in: query
          description: Địa điểm chứa chuỗi này
          schema:
            type: string
            maxLength: 100
        - name: minGoal
          in: query
          description: Mục tiêu vốn tối thiểu (VNĐ)
          schema:
            type: number
            minimum: 0
        - name: maxGoal
          in: query
          description: Mục tiêu vốn tối đa (VNĐ)
          schema:
            type: number
            minimum: 0
        - name: minProgress
          in: query
          description: Tỷ lệ hoàn thành tối thiểu (%)
          schema:
            type: number
            minimum: 0
            maximum: 1000
        - name: maxProgress
          in: query
          description: Tỷ lệ hoàn thành tối đa (%)
          schema:
            type: number
            minimum: 0
            maximum: 1000
        - name: endingWithinDays
          in: query
          description: Chỉ chiến dịch `active` kết thúc trong N ngày tới
          schema:
            type: integer
            minimum: 1
            maximum: 365
        - $ref: "#/components/parameters/CampaignSort"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Danh sách chiến dịch có phân trang
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignList"
        "400":
          description: Trạng thái chưa công khai, tham số sai kiểu/ngoài phạm vi hoặc khoảng lọc ngược
    post:
      operationId: createCampaign
      summary: Tạo chiến dịch mới (trạng thái `draft`)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignRequest"
      responses:
        "201":
          description: Đã tạo chiến dịch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
...
  /donations:
    post:
      operationId: createDonation
      summary: Tạo đơn tài trợ (trạng thái `pending`)
      description: >-
        Idempotent theo `idempotencyKey` — gọi lại với cùng khóa trả về đơn cũ.
        Chặn nếu chiến dịch không ở trạng thái `active` hoặc đã hết hạn.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDonationRequest"
      responses:
        "201":
          description: Đơn tài trợ đã tạo (hoặc đơn cũ nếu trùng idempotency key)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
        "404":
          description: Chiến dịch không tồn tại
        "409":
          description: Chiến dịch không ở trạng thái `active` hoặc đã hết hạn
...
  /admin/overview:
    get:
      operationId: adminOverview
      summary: Số liệu tổng quan cho dashboard admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tổng quan hệ thống
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverview"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/campaigns:
    get:
      operationId: adminListCampaigns
      summary: Danh sách chiến dịch cho admin (mọi trạng thái, kèm chủ dự án)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CampaignQ"
        - $ref: "#/components/parameters/CampaignStatusFilter"
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, ending, raised, backers]
            default: newest
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Danh sách chiến dịch (admin)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignList"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/donations:
    get:
      operationId: adminListDonations
      summary: Danh sách giao dịch cho admin (kèm người tài trợ và chiến dịch)
      description: >-
        Mỗi giao dịch kèm `riskLevel` = mức cao nhất trong các cảnh báo rủi ro đang `open`
        của chiến dịch hoặc người ủng hộ liên quan (null nếu không có). Cảnh báo gắn với
        chiến dịch/người dùng, không gắn từng giao dịch.
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/DonationStatus"
        - name: flagged
          in: query
          description: "`true` = chỉ giao dịch của chiến dịch/người ủng hộ đang có cảnh báo `open`"
          schema:
            type: boolean
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Danh sách giao dịch (admin)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationList"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users:
    get:
      operationId: adminListUsers
      summary: Danh sách người dùng cho admin (lọc/tìm kiếm)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CampaignQ"
        - name: role
          in: query
          schema:
            type: string
            enum: [user, campaign_owner, admin]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, banned]
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Danh sách người dùng (admin)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/audit-logs:
    get:
      operationId: adminListAuditLogs
      summary: Xem nhật ký kiểm toán (append-only)
      security:
        - bearerAuth: []
      parameters:
        - name: action
          in: query
          schema:
            type: string
            maxLength: 100
        - name: entity
          in: query
          schema:
            type: string
            maxLength: 100
        - name: entityId
          in: query
          schema:
            type: string
            maxLength: 36
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Danh sách bản ghi audit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogList"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/risks:
    get:
      operationId: adminListRisks
      summary: Danh sách cảnh báo rủi ro AI
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved, dismissed]
        - name: level
          in: query
          schema:
            type: string
            enum: [low, medium, high]
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Danh sách cảnh báo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAlertList"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/risks/generate:
    post:
      operationId: adminGenerateRisks
      summary: Quét lại cảnh báo rủi ro cho các chiến dịch đang hoạt động
      description: >-
        Gọi AI fraud detection; nếu AI service lỗi thì tự rơi về bộ luật dự
        phòng (`aiAvailable: false`) thay vì báo lỗi cho admin.
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Kết quả quét
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateRisksResult"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/risks/{id}/status:
    patch:
      operationId: adminUpdateRiskStatus
      summary: Xác nhận đã xử lý hoặc bỏ qua một cảnh báo (human-in-the-loop)
      description: Ghi audit `risk_alert.status.update`, lưu `resolvedBy`/`resolvedAt`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRiskStatusRequest"
      responses:
        "200":
          description: Đã cập nhật cảnh báo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAlert"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/campaigns/{id}/featured:
    patch:
      operationId: adminSetCampaignFeatured
      summary: Chọn/bỏ dự án nổi bật ở trang chủ (chỉ admin)
      description: >-
        Chỉ chiến dịch `approved/active/success`; tối đa 6 dự án nổi bật cùng lúc.
        Ghi audit `campaign.feature` / `campaign.unfeature`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [featured]
              properties:
                featured: { type: boolean }
      responses:
        "200":
          description: Chiến dịch sau khi cập nhật
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          description: Chiến dịch chưa công khai
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Đã đủ 6 dự án nổi bật
  /admin/exports/campaigns:
    get:
      operationId: adminExportCampaigns
      summary: Xuất CSV thống kê theo chiến dịch (không dữ liệu cá nhân)
      description: >-
        Mục tiêu, đã huy động, tỷ lệ đạt, lượt ủng hộ/xem, số mốc/hoàn thành, chi đã báo cáo.
        Chủ dự án giả danh `CDA-…` (SHA-256), không tên/email. Chống CSV injection. Audit `admin.export`.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: File CSV UTF-8 (có BOM)
          content:
            text/csv:
              schema:
                type: string
        "401":
          description: Chưa đăng nhập
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/exports/donations-daily:
    get:
      operationId: adminExportDailyDonations
      summary: Xuất CSV ủng hộ theo ngày (giờ Việt Nam) × lĩnh vực
      description: Số lượt/tiền thành công và đã hoàn; mặc định 30 ngày gần nhất. Audit `admin.export`.
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: File CSV UTF-8 (có BOM)
          content:
            text/csv:
              schema:
                type: string
        "401":
          description: Chưa đăng nhập
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /admin/users/{id}/status:
    patch:
      operationId: adminUpdateUserStatus
      summary: Khóa/mở khóa tài khoản người dùng
      description: >-
        Admin không tự khóa chính mình. Ghi audit `user.status.update`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserStatusRequest"
      responses:
        "200":
          description: Đã cập nhật trạng thái tài khoản
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Admin cố tự khóa chính mình
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CampaignQ:
      name: q
      in: query
      description: Tìm kiếm toàn văn (tiêu đề/mô tả/địa điểm hoặc tên/email tùy endpoint)
      schema:
        type: string
        maxLength: 200
    CampaignCategory:
      name: category
      in: query
      schema:
        type: string
        maxLength: 100
    CampaignStatusFilter:
      name: status
      in: query
      description: Một hoặc nhiều trạng thái, cách nhau bởi dấu phẩy (vd. `active,success`)
      style: form
      explode: false
      schema:
        type: array
        items:
          $ref: "#/components/schemas/CampaignStatus"
    CampaignSort:
      name: sort
      in: query
      schema:
        type: string
        enum: [featured, popular, ending, newest, progress, latest]
        default: popular
      description: "`featured`: dự án admin chọn nổi bật trước (mới chọn trước), còn lại theo độ phổ biến"
  responses:
    Forbidden:
      description: Không có quyền thực hiện hành động này
    NotFound:
      description: Không tìm thấy tài nguyên
...
  schemas:
...
    UserSummary:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, campaign_owner, admin]
        emailVerified:
          type: boolean
          description: Có trong phản hồi đăng nhập/đăng ký/refresh/đổi mật khẩu
        aiTrackingConsent:
          type: boolean
          description: Đã đồng ý ghi nhận hành vi cho gợi ý AI
        aiConsentDecided:
          type: boolean
          description: false = chưa từng chọn → frontend hiển thị lời mời đồng ý
    User:
      type: object
      description: >-
        Không bao giờ chứa passwordHash/failedLoginCount/lockedUntil. `email` và `phone`
        chỉ có ở API của chính chủ (/users/*) và admin (/admin/*); bị ẩn khi user xuất hiện
        lồng trong API công khai (vd. `owner` của chiến dịch).
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, campaign_owner, admin]
        status:
          type: string
          enum: [active, inactive, banned]
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        organization:
          type: string
          nullable: true
        emailVerified:
          type: boolean
        aiTrackingConsent:
          type: boolean
          description: Chỉ có ở API của chính chủ/admin
        aiConsentUpdatedAt:
          type: string
          format: date-time
          nullable: true
          description: NULL = chưa từng chọn; chỉ có ở API của chính chủ/admin
        createdAt:
          type: string
          format: date-time
...
    UpdateUserStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, inactive, banned]
    UserList:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/User"
    PageInfo:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    CampaignStatus:
      type: string
      enum:
        - draft
        - pending
        - approved
        - rejected
        - needs_info
        - active
        - paused
        - success
        - failed
        - cancelled
        - ended
...
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        category:
          type: string
        ownerId:
          type: string
          format: uuid
        owner:
          nullable: true
          $ref: "#/components/schemas/UserSummary"
        goalAmount:
          type: number
        currentAmount:
          type: number
          description: Sinh từ tổng giao dịch `completed`, không nhận nhập tay
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/CampaignStatus"
        isFeatured:
          type: boolean
          description: Admin chọn hiển thị ở mục "Dự án nổi bật" trang chủ
        imageUrl:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        backerCount:
          type: integer
        viewCount:
          type: integer
        rejectionReason:
          type: string
          nullable: true
          description: Dùng chung cho lý do từ chối, yêu cầu bổ sung hoặc tạm dừng
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCampaignRequest:
      type: object
      required: [title, description, category, goalAmount, endDate]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        category:
          type: string
          maxLength: 100
        goalAmount:
          type: number
          minimum: 1000
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        imageUrl:
          type: string
          format: uri
        location:
          type: string
          maxLength: 255
...
    CampaignList:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Campaign"
    DonationStatus:
      type: string
      description: >-
        `expired` = đơn `pending` quá `DONATION_PENDING_TTL_MINUTES` phút (mặc định 30) do job
        10 phút/lần chuyển; `cancelled` = người ủng hộ hủy hoặc cổng báo hủy.
      enum: [pending, completed, failed, refunded, expired, cancelled]
    Donation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        campaignId:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          type: string
          example: VND
        status:
          $ref: "#/components/schemas/DonationStatus"
        paymentMethod:
          type: string
          nullable: true
        transactionId:
          type: string
          nullable: true
        idempotencyKey:
          type: string
        message:
          type: string
          nullable: true
        isAnonymous:
          type: boolean
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateDonationRequest:
      type: object
      required: [campaignId, amount, paymentMethod, idempotencyKey]
      properties:
        campaignId:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 20000
        paymentMethod:
          type: string
          enum: [wallet, payos]
        message:
          type: string
          maxLength: 500
        isAnonymous:
          type: boolean
          default: false
        idempotencyKey:
          type: string
          maxLength: 255
        rewardTierId:
          type: string
          format: uuid
          description: >-
            Mức quà chọn kèm. Số tiền phải ≥ `minAmount` (400), mức phải đang nhận và còn suất
            (409). Suất chỉ bị trừ khi thanh toán được xác nhận; nếu hết suất đúng lúc đó khoản
            ủng hộ vẫn được ghi nhận nhưng không kèm quà (`rewardTierId` = null).
...
    DonationList:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Donation"
...
    AdminOverview:
      type: object
      properties:
        users:
          type: object
          properties:
            total: { type: integer }
            active: { type: integer }
            banned: { type: integer }
            owners: { type: integer }
            newThisMonth: { type: integer }
        campaigns:
          type: object
          properties:
            total: { type: integer }
            draft: { type: integer }
            pending: { type: integer }
            needsInfo: { type: integer }
            approved: { type: integer }
            active: { type: integer }
            ended: { type: integer }
            success: { type: integer }
            failed: { type: integer }
            rejected: { type: integer }
        donations:
          type: object
          properties:
            total: { type: integer }
            completed: { type: integer }
            failed: { type: integer }
            pending: { type: integer }
            refunded: { type: integer }
            completedAmount: { type: number }
            failedAmount: { type: number }
            expired: { type: integer }
            cancelled: { type: integer }
        successRate:
          type: number
          description: Tỷ lệ `success / (success + failed)` trong các chiến dịch đã kết thúc
        risks:
          type: object
          properties:
            open: { type: integer }
            high: { type: integer }
            medium: { type: integer }
        topCategories:
          type: array
          items:
            type: object
            properties:
              category: { type: string }
              campaigns: { type: integer }
              raised: { type: number }
        monthly:
          type: array
          description: 6 tháng gần nhất, tính đến tháng hiện tại
          items:
            type: object
            properties:
              month:
                type: string
                example: "2026-09"
              raised: { type: number }
              donations: { type: integer }
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          nullable: true
          description: Bỏ trống khi hành động do hệ thống/webhook thực hiện
        action:
          type: string
        entity:
          type: string
        entityId:
          type: string
          nullable: true
        oldValues:
          type: object
          nullable: true
          additionalProperties: true
        newValues:
          type: object
          nullable: true
          additionalProperties: true
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    AuditLogList:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AuditLog"
    RiskReason:
      type: object
      properties:
        group:
          type: string
        label:
          type: string
        weight:
          type: number
    RiskAlert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entityType:
          type: string
          enum: [campaign, user]
        entityId:
          type: string
          format: uuid
        entityName:
          type: string
        level:
          type: string
          enum: [low, medium, high]
        score:
          type: number
        method:
          type: string
          enum: [ai, rule]
        reasons:
          type: array
          items:
            $ref: "#/components/schemas/RiskReason"
        evidences:
          type: object
          additionalProperties:
            type: number
        status:
          type: string
          enum: [open, resolved, dismissed]
        resolvedBy:
          type: string
          format: uuid
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    RiskAlertList:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/RiskAlert"
    UpdateRiskStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [open, resolved, dismissed]
    GenerateRisksResult:
      type: object
      required: [aiAvailable, scanned, flagged]
      properties:
        aiAvailable:
          type: boolean
          description: "false nghĩa là đã dùng bộ luật dự phòng thay vì mô hình AI"
        scanned:
          type: integer
        flagged:
          type: integer
...